+++ tests/test_arrow.py
@@ -2266,6 +2266,25 @@ class TestArrowHumanize:
         with pytest.raises(ValueError):
             arw.humanize(later, granularity="week")
 
+    # Bulgarian is an example of a language that overrides _format_timeframe
+    # Applicabale to all locales. Note: Contributors need to make sure
+    # that if they override describe or describe_mutli, that delta
+    # is truncated on call
+
+    def test_no_floats(self):
+        arw = arrow.Arrow(2013, 1, 1, 0, 0, 0)
+        later = arw.shift(seconds=55000)
+        humanize_string = arw.humanize(later, locale="bg", granularity="minute")
+        assert humanize_string == "916 минути назад"
+
+    def test_no_floats_multi_gran(self):
+        arw = arrow.Arrow(2013, 1, 1, 0, 0, 0)
+        later = arw.shift(seconds=55000)
+        humanize_string = arw.humanize(
+            later, locale="bg", granularity=["second", "minute"]
+        )
+        assert humanize_string == "916 минути 40 няколко секунди назад"
+
 
 @pytest.mark.usefixtures("time_2013_01_01")
 class TestArrowHumanizeTestsWithLocale:
@@ -2382,6 +2401,15 @@ def locale_list_no_weeks():
         "ms-bn",
         "or",
         "or-in",
+        "se",
+        "se-fi",
+        "se-no",
+        "se-se",
+        "lb",
+        "lb-lu",
+        "ta",
+        "ta-in",
+        "ta-lk",
     ]
 
     return tested_langs
@@ -2437,6 +2465,11 @@ def locale_list_with_weeks():
         "ms",
         "ms-my",
         "ms-bn",
+        "lb",
+        "lb-lu",
+        "ta",
+        "ta-in",
+        "ta-lk",
     ]
 
     return tested_langs